fix WorkManager
