+++ app/build.gradle
@@ -11,7 +11,7 @@ ext {
     dagger_version = "2.16"
     lifecycle_version = "1.1.1"
     coroutines_version = "0.24.0"
-    work_runtime_version = "1.0.0-alpha06"
+    work_runtime_version = "1.0.0-alpha04"
 }
 
 android {
